Update Arrays are fun.cpp
Problem Statement:

Arrays are really fun to deal with. All the array related problems are not always hard, some problems just need to be looked from a different angle. This is one of the similar problem. You will be given an array A of length n. 
Find the sum of minimum number in every subarray S of A.

A subarray is part of the array between index i and j such that 0<i<=j<n